extract hook and gridlines component
