+++ frontend/src/cards/RateBarChartCard.tsx
@@ -1,5 +1,5 @@
+import { addComparisonAllsRowToIntersectionalData } from '../charts/rateBarChart/helpers'
 import { RateBarChart } from '../charts/rateBarChart/Index'
-import { addComparisonAllsRowToIntersectionalData } from '../charts/simpleBarHelperFunctions'
 import { generateChartTitle, generateSubtitle } from '../charts/utils'
 import type { DataTypeConfig, MetricId } from '../data/config/MetricConfigTypes'
 import { isPctType } from '../data/config/MetricConfigUtils'
